linkedin_extractor/requirements.txt
playwright==1.44.0
paddleocr==2.7.3
paddlepaddle==2.6.1
transformers>=4.40.0
torch>=2.2.0
accelerate>=0.29.0
pydantic==2.7.1
rich==13.7.1
python-dotenv==1.0.1
ftfy==6.2.0
loguru==0.7.2
Pillow==10.3.0
